requirements.txt
requests~=2.32.3
beautifulsoup4~=4.13.4
fastapi~=0.115.12
pydantic~=2.11.3
numpy<2.0                  # or numpy~=2.0.2 if required
pytest~=8.3.5
crawler~=0.0.2
faiss-cpu==1.7.4
openai~=0.28.0
ollama~=0.4.8
python-dotenv~=1.1.0